+++ src/isotools/transcriptome.py
@@ -244,7 +244,7 @@ class Transcriptome:
     from ._transcriptome_io import gene_table, transcript_table, chimeric_table, write_gtf, export_alternative_splicing
 
     # filtering functionality and iterators
-    from ._transcriptome_filter import add_qc_metrics, add_filter, remove_filter, iter_genes, iter_transcripts, iter_ref_transcripts
+    from ._transcriptome_filter import add_qc_metrics, add_orf_prediction, add_filter, remove_filter, iter_genes, iter_transcripts, iter_ref_transcripts
 
     # statistic: differential splicing, alternative_splicing_events
     from ._transcriptome_stats import die_test, altsplice_test, coordination_test, alternative_splicing_events, rarefaction